infra: use Go package path names for providers
Summary:
This changes infra.Register to compute a provider name based on the
type's package path and name, following Go convention (path.name).
This way, it's always clear exactly where a provider is coming from,
and we lean on Go's package namespace for uniqueness. It also helps
us with clearer error messages as a provider is always linked in by
its package path.

Reviewers: pgopal, smahadevan, jcharumilind

Reviewed By: pgopal, smahadevan

Differential Revision: https://phabricator.grailbio.com/D23367

fbshipit-source-id: 9619702